working delete with modal pop up

+++ client/src/components/RecipeCard.jsx
@@ -1,7 +1,7 @@
 'use client' // client component, not server rendered
 import { useState } from 'react';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import { Card, Col, Row, Container, CardFooter } from 'react-bootstrap';
+import { Card, Col, Row, Container, Modal, Button } from 'react-bootstrap';
 import PlaylistAddIcon from '@mui/icons-material/PlaylistAdd';
 import EditNoteIcon from '@mui/icons-material/EditNote';
 import DeleteForeverIcon from '@mui/icons-material/DeleteForever';
@@ -31,6 +31,17 @@ const faunaOne = Fauna_One({
 //Recipe card takes delete function as a prop which is actioned on the delete icon press
 export default function RecipeCard({ recipes, baseURL }) {
     const [currentRecipes, setCurrentRecipes] = useState(recipes);
+    const [showModal, setShowModal] = useState(false);
+    const [recipeToDelete, setRecipeToDelete] = useState(null);
+    const [recipeTitle, setRecipeTitle] = useState(null);
+
+    //handle the modal popup
+    const handleCloseModal = () => setShowModal(false);
+    const handleShowModal = (recipeId, recipeTitle) => {
+        setRecipeToDelete(recipeId);
+        setRecipeTitle(recipeTitle);
+        setShowModal(true);
+    }
 
     //add this function to refetch the data from the server once a delete/update/add has been performed
     const fetchRecipes = async () => {
@@ -43,24 +54,29 @@ export default function RecipeCard({ recipes, baseURL }) {
         }
     };
 
-    const deleteRecipe = async (recipeId) => {
-        try {
-            const response = await fetch(`${baseURL}/api/recipes`, {
-                method: 'DELETE',
-                headers: {
-                    'Content-Type': 'application/json'
-                },
-                body: JSON.stringify({ recipe_id: recipeId })
-            });
+    const deleteRecipe = async () => {
+        if (recipeToDelete) {
+            try {
+                const response = await fetch(`${baseURL}/api/recipes`, {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json'
+                    },
+                    body: JSON.stringify({ recipe_id: recipeToDelete })
+                });
 
-            if (response.ok) {
-                //setCurrentRecipes(prevRecipes => prevRecipes.filter(recipe => recipe.recipe_id !== recipeId));
-                fetchRecipes();
-            } else {
-                console.error('Failed to delete recipe');
+                if (response.ok) {
+                    fetchRecipes();
+                } else {
+                    console.error('Failed to delete recipe');
+                }
             }
-        } catch (error) {
-            console.error('Error deleting recipe:', error);
+            catch (error) {
+                console.error('Error deleting recipe:', error);
+            }
+            finally {
+                handleCloseModal();
+            }
         }
     };
 
@@ -82,36 +98,54 @@ export default function RecipeCard({ recipes, baseURL }) {
     }
 
     return (
-        <Container className='justify-content-center align-items-center'>
-            <Row xs={1} sm={2} md={5} className="justify-content-center">
-                {currentRecipes.map(recipe => (
-                    <Col key={recipe.recipe_id} className="g-3 justify-content-center">
-                        <Card className=' border-2 text-center'>
-                            <Card.Img
-                                className="mx-auto pt-2"
-                                src={recipe.image ? recipe.image : "/images/plate.png"}
-                                alt={recipe.recipe_title}
-                            />
-                            <Card.Body>
-                                <Card.Title className={`${fontCinzel.className} title center fs-5`}>{recipe.recipe_title}</Card.Title>
-                                <Tooltip title="Add to Meal Plan" arrow>
-                                    <a className="btn btn-link btn-floating btn-outline-dark btn-lg text-dark icon-button" href="#!" role="button">
-                                        <PlaylistAddIcon className='custom-icon' /></a>
-                                </Tooltip>
-                                <Tooltip title="Edit Recipe" arrow>
-                                    <a className="btn btn-link btn-floating btn-outline-dark btn-lg text-dark icon-button" href="#!" role="button">
-                                        <EditNoteIcon className='custom-icon' /></a>
-                                </Tooltip>
-                                <Tooltip title="Delete Recipe" arrow>
-                                    <a className="btn btn-link btn-floating btn-outline-dark btn-lg text-dark icon-button" href="#!" role="button"
-                                        onClick={() => deleteRecipe(recipe.recipe_id)}>
-                                        <DeleteForeverIcon className='custom-icon' /></a>
-                                </Tooltip>
-                            </Card.Body>
-                        </Card>
-                    </Col>
-                ))}
-            </Row>
-        </Container >
+        <div>
+            <Container className='justify-content-center align-items-center'>
+                <Row xs={1} sm={2} md={5} className="justify-content-center">
+                    {currentRecipes.map(recipe => (
+                        <Col key={recipe.recipe_id} className="g-3 justify-content-center">
+                            <Card className=' border-2 text-center'>
+                                <Card.Img
+                                    className="mx-auto pt-2"
+                                    src={recipe.image ? recipe.image : "/images/plate.png"}
+                                    alt={recipe.recipe_title}
+                                />
+                                <Card.Body>
+                                    <Card.Title className={`${fontCinzel.className} title center fs-5`}>{recipe.recipe_title}</Card.Title>
+                                    <Tooltip title="Add to Meal Plan" arrow>
+                                        <a className="btn btn-link btn-floating btn-outline-dark btn-lg text-dark icon-button" href="#!" role="button">
+                                            <PlaylistAddIcon className='custom-icon' /></a>
+                                    </Tooltip>
+                                    <Tooltip title="Edit Recipe" arrow>
+                                        <a className="btn btn-link btn-floating btn-outline-dark btn-lg text-dark icon-button" href="#!" role="button">
+                                            <EditNoteIcon className='custom-icon' /></a>
+                                    </Tooltip>
+                                    <Tooltip title="Delete Recipe" arrow>
+                                        <a className="btn btn-link btn-floating btn-outline-dark btn-lg text-dark icon-button" href="#!" role="button"
+                                            onClick={() => handleShowModal(recipe.recipe_id, recipe.recipe_title)}>
+                                            <DeleteForeverIcon className='custom-icon' /></a>
+                                    </Tooltip>
+                                </Card.Body>
+                            </Card>
+                        </Col>
+                    ))}
+                </Row>
+            </Container >
+
+            {/*Modal Pop Up */}
+            <Modal show={showModal} onHide={handleCloseModal}>
+                <Modal.Header closeButton>
+                    <Modal.Title>Delete {recipeTitle}?</Modal.Title>
+                </Modal.Header>
+                <Modal.Body>Are you sure you want to delete this recipe?</Modal.Body>
+                <Modal.Footer>
+                    <Button variant="secondary" onClick={handleCloseModal}>
+                        Cancel
+                    </Button>
+                    <Button variant="danger" onClick={deleteRecipe}>
+                        Delete
+                    </Button>
+                </Modal.Footer>
+            </Modal>
+        </div>
     )
 }
